Fixing the build on the warnings_as_errors buildbot slave.  To do this, I:
Added explicit cast from double to Size in: 
   core/scoring/electron_density_atomwise/ElectronDensityAtomwise.cc 
   protocols/swa/rna/StepWiseRNA_Base_Sugar_Rotamer.cc
   protocols/swa/rna/StepWiseRNA_ResidueSampler.cc

Added explicit cast from double to int in:
   protocols/swa/rna/StepWiseRNA_FloatingBase_Sampler_Util.cc
	 protocols/swa/rna/StepWiseRNA_ResidueSampler.cc
   protocols/swa/rna/StepWiseRNA_VDW_Bin_Screener.cc
												  


git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@47409 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 7cd7fb878838e07fea6647b63dfe64d97199318c

+++ rosetta_source/src/protocols/swa/rna/StepWiseRNA_FloatingBase_Sampler_Util.cc
@@ -357,13 +357,13 @@ namespace rna {
 //		using namespace Bin_size;
 
 		Base_bin base_bin;
-		base_bin.centroid_x=centriod[0]/centroid_bin_size; 
-		base_bin.centroid_y=centriod[1]/centroid_bin_size; 
-		base_bin.centroid_z=centriod[2]/centroid_bin_size; 		
+		base_bin.centroid_x=int(centriod[0]/centroid_bin_size); 
+		base_bin.centroid_y=int(centriod[1]/centroid_bin_size); 
+		base_bin.centroid_z=int(centriod[2]/centroid_bin_size); 		
 
-		base_bin.euler_alpha=euler_angles.alpha/euler_angle_bin_size;
-		base_bin.euler_gamma=euler_angles.gamma/euler_angle_bin_size;
-		base_bin.euler_z=euler_angles.z/euler_z_bin_size;
+		base_bin.euler_alpha=int(euler_angles.alpha/euler_angle_bin_size);
+		base_bin.euler_gamma=int(euler_angles.gamma/euler_angle_bin_size);
+		base_bin.euler_z=int(euler_angles.z/euler_z_bin_size);
 
 
 		if(centriod[0]<0) base_bin.centroid_x--;
